Finish Day 16
I finished Part 1 relatively fast on the day placing #237 in 34 minutes and 19 seconds.
However, this was mostly luck, and I had to write and tear down multiple solutions along with asking for guidance to be able to finish Part 2.
This final solution is optimized based on my findings and runs in under 500ms on my input, which is pretty good.