Finish Day 18
I did not place today, but I was surprised by the ease of Part 1.
I did Part 1 in 3 minutes and 39 seconds, and that was only good enough for #196.
Part 2 was a bit harder and after the leaderboard filled up, I took the time to actually make a ZDirection helper class rather than use 3 yucky loops.
Interestingly, running a DFS in each potential steam/air pocket seems to be faster than running a DFS from the outside by expanding the bounds by 1 in each direction by about 10 ms on my hardware.
Logically, this would make sense due to the fact that you have to check a minimum of 3176 extra points in the latter way as that extra layer will always be points of steam.
It appears to be 3176 points because the volume of the original cube is 22^3, so you have to subtract that from the volume of the larger encompassing cube which is 24^3, giving 24^3-22^3=3176.