Finish Day 19
I did not place on this day.
I gave up trying to find the heuristic myself and cheated a bit.
The rest of the code outside of the strategy is my own.
I originally attempted the puzzle in Kotlin but switched to Java because of how bad my Kotlin code was, and I don't feel like porting it back to Kotlin again right now.