Finish Day 22
I did not find this day extremely challenging, but visualizing Part 2 without a real-life cube was hard.
For Part 2, I wrote the cube mappings based on the example input assuming it would be the same as the real input.
This mistake cost me time and sanity.
My code has the cube mappings for the example input and real input, but you have to manually swap between the two in the code.
I tried to write an auto-folder, but it failed horribly.